compile: intro of option files for modules with special needs
This allows the user to specify compiler arguments for each module with files.
To do so for a given module "json", the user inputs the args for the compiler
in a file named json.nit.args. As example, the content of the file could be
"--cc-lib-name json", which asks to link with a native library when using this
module. Theses args will be interpreted as those given on the command line.

The args allowed are limited to C compilation options, as others don't seam
to be of any use in this way.

Signed-off-by: Alexis Laferrière <alexis.laf@xymus.net>