model: new metamodel
This commit introduces a new metamodel.
I expect it is simpler than the current one.

A new modelbuilder and typing is also provided.
As a proof of concept two tools are included: nit and nitstats

nit is a basic naive and inefficient interpreter.
nitstats is a tool to collect and gather some statistics.

Signed-off-by: Jean Privat <jean@pryen.org>