ni: cleans up code repetition in native interface and improves performance
The code clean up uses trasitionnal variables (NiVariable) to simplyfy the
concept of variable transitionning from C to Nit and Nit to C. It allows
code factorisation and cleaner algorithms.

An extern method without explicit callback and not returning a complexe Nit
object cannot legally use any references. Knowing this, we do not generate
full local references for the arguments of such methods. This avoids to call
malloc and free on unused data structure and greatly reduces overhead for
simple extern methods.

Signed-off-by: Alexis Laferrière <alexis.laf@xymus.net>