nitg: ad hoc separate compiler stub module
separate_compiler refines global_compiler to add a simulated separate
compilation process.
A new option --separate is also added to nitg.

Currently this does not even generate a valid C code for
examples/hello_world.nit but C for ests/base_simple3.nit is valid
(but nonfunctional).

Many parts are missing, especially the computation of tables and the
implementation of the main OO mechanisms.

The short therm usage of this code is to help the evaluation of subtyping
implementations of OO language.
Eventually, the code could be the base of a cleaner naive separate
compiler (in order to replace the current nitc).

Signed-off-by: Jean Privat <jean@pryen.org>