model: remove a lot of casts to MGenericType
Since the generalization of MGenericType::arguments,
MGenericType do not introduce specific methods anymore.
Therefore, there is no point to require specific type or casts in clients.

This simplify the code since now a MClass, a MClassDef and a MClassType
can all be used to designate an entity, generic or not.

Signed-off-by: Jean Privat <jean@pryen.org>