nitg: add type-checks on covariant parameters
The compilation of a MMethodDef will add required type checks before
compiling the body of the method.
The type checks are implemented trough v.type_test so this patch
is compatible with other engines.

problem: most nitg-sep tests become [todo] since virtual type tests
are not implemented is the engine but become mandatory in many methods
of Comparable.

Signed-off-by: Jean Privat <jean@pryen.org>