nitc/ffi: allows to declare C type of extern classes from Nit
This feature uses the FFI grammar but acts at the level of the native
interface. It allows the user to specify the C equivalent of a Nit
extern class. Instances of extern classes are automatically converted
to this type when passed to C. The equivalent C type must be a pointer
or compatible with a pointer type.

Signed-off-by: Alexis Laferrière <alexis.laf@xymus.net>