nitdoc: Fix bug when signing in
Before creating the session, we check if the repository and the branch
informed are existing, whose causing some bugs when one of these do not
exist. If the repository does not exist then we stop the process. If
the branch does not exist we ask to the customer if he wants to create
it.

Signed-off-by: Stefan Lage <lagestfan@gmail.com>