rta: store real types in live_cast_type
The separate compiler need to distinguish if the target is nullable or not.
So do not un-nullable cast types.

rta_metrics.nit is also updated to deal with the change of the static
type of the attribute.

nitg-g is also modified to use the new correct the RTA information
(this remove a workaround)

Signed-off-by: Jean Privat <jean@pryen.org>