syntax: ignore first '\n' in triple-quoted strings
This detail of the specification makes triple-quoted strings more
readable and usable in code.

Before, programmer was required to starts content just after the """
thus writing unaligned piece of text

~~~
var x = """Roses are red,
Violets are blue,
Triple-quoted strings,
Are easier to write.
"""
~~~

This patch ignore the end of line iff it just follows the """,
thus programmer can write

~~~
var x = """
Roses are red,
Violets are blue,
Triple-quoted strings,
Are easier to write.
"""
~~~

This commit will improve the usability of the new/future template library.

Signed-off-by: Jean Privat <jean@pryen.org>