model: fix a compiler crash on access to null constructor signature
Before this patch, compilation process stopped on a null pointer exception.
Because of the compilation workflow, nitg tried to access a super init even
if this init was malformed (type error) and thus not modelized properly.

This commit add one more condition to ensure that msignature is initialized before the test
and raise the correct error instead of crashing.

Signed-off-by: Alexandre Terrasa <alexandre@moz-code.org>