modelize_property: enable the access of the internal attributes
The transition from old-style to new-style attributes is a mess.
So in order to facilitate the transition, this patch enable
the access to the internal attribute previously hidden by then
new_style attribute.

Therefore, the transition path is simpler since one can move declaration
from old-style to new-style but some access can still use the old `_foo`
notation.

Note, the `_foo` access is not mean to be publicized bu only made available
for specific short-term workarounds and transitions.

Practically, internal attribute named are prefixed with `_` and
made visible (but private)