mongodb: fixes crash when the return of `mongoc_cursor_current` was freed
As explained by the documentation (that I should have read more thoroughly...),
`bson_t` structures allocated by the `mongoc_cursor_current` function should not be freed by the user.

Before this commit, the `bson_t` returned by `mongoc_cursor_current` where left as this then freed
by the nit GC.

This commit makes the `current` method to return a copy of the `bson_t` struct so
it can be freed safely by the GC later.

And... this is why I fear and hate the C programming language... :)

Signed-off-by: Alexandre Terrasa <alexandre@moz-code.org>